+++ test/test_boot_not_finised.py
@@ -5,15 +5,13 @@ from .helper import execute_main, MPopen
 class TestBootupNotFinished(unittest.TestCase):
 
     def test_bootup_not_finished(self):
-        result = execute_main(popen=(
+        result = execute_main(argv=['--no-performance-data'], popen=(
             MPopen(stdout='systemctl-list-units_ok.txt'),
             MPopen(returncode=1, stderr='systemd-analyze_not-finished.txt')
         ))
         self.assertEqual(result.exitcode, 0)
         self.assertEqual(
-            'SYSTEMD OK - all '
-            '| count_units=386 units_activating=0 '
-            'units_active=275 units_failed=0 units_inactive=111',
+            'SYSTEMD OK - all',
             result.first_line
         )
 
@@ -27,11 +25,6 @@ class TestBootupNotFinished(unittest.TestCase):
         self.assertEqual(result.exitcode, 0)
         self.assertIn(
             'SYSTEMD OK - all\n',
-            result.output
-        )
-        self.assertIn(
-            '| count_units=386 units_activating=0 '
-            'units_active=275 units_failed=0 units_inactive=111\n',
             result.output
         )